Added update method to books controller and tests
It's now possible to select a book's author from within a select tag in
the create and edit form. I decided against usin nested attributes for
this as I didn't want users to accidentally make changes to author details whilst updating a book.
I was unable to get the form_for tag to work on the edit page without
the use of a hidden input tag which still strikes me as unnecessary.
Though the edit form now works in the development environment, the related test still fails. The create test passes however.