Authors controller and CarrierWave gem added
The Authors controller and views have been added to the app. The views
should now display information stored on the database regarding
individual authors. I have added the CarrierWave gem to the app to allow
users to upload pictures of the author and save it in the public folder.

Nested attributes have been added to the form partial in the authors view
which will allow users to update an author's existing books. However, I
was unable to set up a helper method that would allow users to add new
books to the author form and so have had to work around this with a link
to the new book form.

Many of the tests written for the authors controller are failing and I'm
unsure as to why. I'm also still unable to update books in the test
environment even though I can do this on the front end.