Have attempted to integrate the Twitter API into the app
Have added the Twitter gem to the app but am unable to
connect to the API using my credentials. Details are being stored in
environmental variables with the use of the Figaro gem.
Have also made changes to the index page so that it more closely
resembles the .pdf supplied.